FGQCanvas clipping / default text color fixes
Also fix some compiler warnings, but the main thing is text now
defaults to white correctly if no color was set explicitly, and
local/parent clip-reference-frames are supported.

+++ utils/fgqcanvas/fgcanvaselement.h
@@ -140,7 +140,7 @@ private:
     mutable ReferenceFrame _clipFrame = ReferenceFrame::GLOBAL;
 
     void parseCSSClip(QByteArray value);
-    float parseCSSValue(QByteArray value) const;
+    double parseCSSValue(QByteArray value) const;
 };
 
 using FGCanvasElementVec = std::vector<FGCanvasElement*>;
